+++ packages/x-global/src/index.ts
@@ -5,6 +5,10 @@ export { packageInfo } from './packageInfo';
 
 type GlobalThis = typeof globalThis;
 
+type GlobalNames = keyof typeof window;
+
+type GlobalType<T extends keyof typeof window> = typeof window[T];
+
 function evaluateThis (fn: (code: string) => unknown): GlobalThis {
   return fn('return this') as GlobalThis;
 }
@@ -20,3 +24,15 @@ export const xglobal = (
           ? window
           : evaluateThis(Function)
 );
+
+export function extractGlobal <N extends GlobalNames, T extends GlobalType<N>> (name: N, fallback: unknown): T {
+  return typeof xglobal[name as 'undefined'] === 'undefined'
+    ? fallback as T
+    : xglobal[name as 'undefined'] as unknown as T;
+}
+
+export function exposeGlobal <N extends GlobalNames> (name: N, fallback: unknown): void {
+  if (typeof xglobal[name as 'undefined'] === 'undefined') {
+    xglobal[name as 'undefined'] = fallback as undefined;
+  }
+}