+++ packages/x-textdecoder/src/shim.ts
@@ -1,9 +1,7 @@
 // Copyright 2017-2021 @polkadot/x-textdecoder authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { xglobal } from '@polkadot/x-global';
+import { exposeGlobal } from '@polkadot/x-global';
 import { TextDecoder } from '@polkadot/x-textdecoder';
 
-if (typeof xglobal.TextDecoder !== 'function') {
-  (xglobal as Record<string, unknown>).TextDecoder = TextDecoder;
-}
+exposeGlobal('TextDecoder', TextDecoder);